Redirection into specific channel 4/18
still many to go

+++ Commands/clear.js
@@ -2,6 +2,7 @@ exports.run = (client, message, args) => {
     if(!message.member.permissions.has("MANAGE_MESSAGES")) return message.channel.send("Sorry, but it seems that you do not have the lowest permission to use this command.")
 
     let amount = parseInt(args.join(" "), 10)
+    if(amount >= 100) return message.channel.send("That was too much! Please try again with a smaller number. _(:_」∠)_")
     if(!amount) amount = 5
     message.channel.bulkDelete(amount+1);
